Move import_object to its own module
